Backend/requirements.txt
fastapi[all]
pydantic
bcrypt
uvicorn
python-multipart
pytest
pytest-mock
python-jose[cryptography]
httpx